+++ apps/preview/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @ensembleui/react-preview
 
+## 0.0.14
+
+### Patch Changes
+
+- b93fc80: add showDialog and closeAllDialogs api
+- Updated dependencies [ddcc772]
+- Updated dependencies [b93fc80]
+  - @ensembleui/react-runtime@0.1.25
+
+## 0.0.13
+
+### Patch Changes
+
+- d789ccb: Update preview for studio integration
+- Updated dependencies [044367a]
+- Updated dependencies [bca1ecf]
+- Updated dependencies [c63ab76]
+- Updated dependencies [54702bc]
+- Updated dependencies [d789ccb]
+  - @ensembleui/react-runtime@0.1.24
+  - @ensembleui/react-framework@0.1.19
+
 ## 0.0.12
 
 ### Patch Changes